Refactoring
Split the mutation and query files into two parts, based on clubs and members. And did some other minor refactoring changes
