+++ mutations_members.py
@@ -2,260 +2,13 @@ import strawberry
 
 from fastapi.encoders import jsonable_encoder
 from datetime import datetime
-import requests
 
-from db import db
+from db import membersdb
 
 # import all models and types
 from otypes import Info
-from models import Club, Member
-from otypes import (
-    FullClubInput,
-    SimpleClubInput,
-    SimpleClubType,
-    FullClubType,
-)
+from models import Member
 from otypes import FullMemberInput, SimpleMemberInput, MemberType
-
-"""
-CLUB MUTATIONS
-"""
-
-
-def updateRole(uid, cookies=None, role="club"):
-    """
-    Function to call the updateRole mutation
-    """
-    try:
-        query = """
-                    mutation UpdateRole($roleInput: RoleInput!) {
-                        updateRole(roleInput: $roleInput)
-                    }
-                """
-        variables = {
-            "roleInput": {
-                "role": role,
-                "uid": uid,
-            }
-        }
-        if cookies:
-            result = requests.post(
-                "http://gateway/graphql",
-                json={"query": query, "variables": variables},
-                cookies=cookies,
-            )
-        else:
-            result = requests.post(
-                "http://gateway/graphql", json={"query": query, "variables": variables}
-            )
-
-        return result.json()
-    except Exception:
-        return None
-
-
-@strawberry.mutation
-def createClub(clubInput: FullClubInput, info: Info) -> SimpleClubType:
-    """
-    Create a new Club.
-    Checks for the 'cc' role, else raises an Error.
-    Checks for uniqueness of the club code.
-    """
-    user = info.context.user
-    if user is None:
-        raise Exception("Not Authenticated")
-
-    role = user["role"]
-    club_input = jsonable_encoder(clubInput.to_pydantic())
-
-    if role in ["cc"]:
-        club_input["cid"] = club_input["email"].split("@")[0]
-
-        cid_exists = db.clubs.find_one({"cid": club_input["cid"]})
-        if cid_exists:
-            raise Exception("A club with this cid already exists")
-
-        code_exists = db.clubs.find_one({"code": club_input["code"]})
-        if code_exists:
-            raise Exception("A club with this short code already exists")
-
-        created_id = db.clubs.insert_one(club_input).inserted_id
-        created_sample = Club.parse_obj(db.clubs.find_one({"_id": created_id}))
-
-        updateRole(club_input["cid"], info.context.cookies)
-
-        return SimpleClubType.from_pydantic(created_sample)
-
-    else:
-        raise Exception("Not Authenticated to access this API")
-
-
-@strawberry.mutation
-def editClub(clubInput: FullClubInput, info: Info) -> FullClubType:
-    """
-    Mutation for editing of the club details either by that specific club or the cc.
-    """
-    user = info.context.user
-    if user is None:
-        raise Exception("Not Authenticated")
-
-    role = user["role"]
-    uid = user["uid"]
-
-    club_input = jsonable_encoder(clubInput.to_pydantic())
-
-    if role in ["cc"]:
-        exists = db.clubs.find_one({"code": club_input["code"]})
-        if not exists:
-            raise Exception("A club with this code doesn't exist")
-        if not exists:
-            raise Exception("A club with this code doesn't exist")
-
-        club_input["state"] = exists["state"]
-        club_input["_id"] = exists["_id"]
-
-        db.clubs.replace_one({"code": club_input["code"]}, club_input)
-        if "socials" in club_input.keys():
-            db.clubs.update_one(
-                {"code": club_input["code"]},
-                {
-                    "$set": {
-                        "socials.website": club_input["socials"]["website"],
-                        "socials.instagram": club_input["socials"]["instagram"],
-                        "socials.facebook": club_input["socials"]["facebook"],
-                        "socials.youtube": club_input["socials"]["youtube"],
-                        "socials.twitter": club_input["socials"]["twitter"],
-                        "socials.linkedin": club_input["socials"]["linkedin"],
-                        "socials.discord": club_input["socials"]["discord"],
-                        "socials.other_links": club_input["socials"]["other_links"],
-                    }
-                },
-            )
-        db.clubs.update_one(
-            {"code": club_input["code"]},
-            {
-                "$set": {
-                    "created_time": exists["created_time"],
-                    "updated_time": datetime.utcnow(),
-                }
-            },
-        )
-
-        if exists["cid"] != club_input["cid"]:
-            updateRole(exists["cid"], info.context.cookies, role="public")
-            updateRole(club_input["cid"], info.context.cookies, role="club")
-
-        result = Club.parse_obj(db.clubs.find_one({"code": club_input["code"]}))
-        return FullClubType.from_pydantic(result)
-
-    elif role in ["club"]:
-        exists = db.clubs.find_one({"cid": club_input["cid"]})
-        if not exists:
-            raise Exception("A club with this cid doesn't exist")
-        if uid != club_input["cid"]:
-            raise Exception("Authentication Error! (CID CHANGED)")
-
-        if (
-            club_input["name"] != exists["name"]
-            or club_input["email"] != exists["email"]
-        ):
-            raise Exception(
-                "You don't have permission to change the name/email of the club. Please contact CC for it"
-            )
-
-        if club_input["category"] != exists["category"]:
-            raise Exception("Only CC is allowed to change the category of club.")
-
-        club_input["state"] = exists["state"]
-        club_input["_id"] = exists["_id"]
-
-        db.clubs.replace_one({"cid": uid}, club_input)
-        if "socials" in club_input.keys():
-            db.clubs.update_one(
-                {"cid": club_input["cid"]},
-                {
-                    "$set": {
-                        "socials.website": club_input["socials"]["website"],
-                        "socials.instagram": club_input["socials"]["instagram"],
-                        "socials.facebook": club_input["socials"]["facebook"],
-                        "socials.youtube": club_input["socials"]["youtube"],
-                        "socials.twitter": club_input["socials"]["twitter"],
-                        "socials.linkedin": club_input["socials"]["linkedin"],
-                        "socials.discord": club_input["socials"]["discord"],
-                        "socials.other_links": club_input["socials"]["other_links"],
-                    }
-                },
-            )
-        db.clubs.update_one(
-            {"cid": club_input["cid"]},
-            {
-                "$set": {
-                    "created_time": exists["created_time"],
-                    "updated_time": datetime.utcnow(),
-                }
-            },
-        )
-
-        result = Club.parse_obj(db.clubs.find_one({"cid": club_input["cid"]}))
-        return FullClubType.from_pydantic(result)
-
-    else:
-        raise Exception("Not Authenticated to access this API")
-
-
-@strawberry.mutation
-def deleteClub(clubInput: SimpleClubInput, info: Info) -> SimpleClubType:
-    """
-    Mutation for the cc to move a club to deleted state.
-    """
-    user = info.context.user
-    if user is None:
-        raise Exception("Not Authenticated")
-
-    role = user["role"]
-    club_input = jsonable_encoder(clubInput)
-
-    if role not in ["cc"]:
-        raise Exception("Not Authenticated to access this API")
-
-    db.clubs.update_one(
-        {"cid": club_input["cid"]},
-        {"$set": {"state": "deleted", "updated_time": datetime.utcnow()}},
-    )
-
-    updateRole(club_input["cid"], info.context.cookies, "public")
-
-    updated_sample = Club.parse_obj(db.clubs.find_one({"cid": club_input["cid"]}))
-
-    return SimpleClubType.from_pydantic(updated_sample)
-
-
-@strawberry.mutation
-def restartClub(clubInput: SimpleClubInput, info: Info) -> SimpleClubType:
-    """
-    Mutation for cc to move a club from deleted state to asctive state.
-    """
-    user = info.context.user
-    if user is None:
-        raise Exception("Not Authenticated")
-
-    role = user["role"]
-    club_input = jsonable_encoder(clubInput)
-
-    if role not in ["cc"]:
-        raise Exception("Not Authenticated to access this API")
-
-    db.clubs.update_one(
-        {"cid": club_input["cid"]},
-        {"$set": {"state": "active", "updated_time": datetime.utcnow()}},
-    )
-
-    updateRole(club_input["cid"], info.context.cookies, "club")
-
-    updated_sample = Club.parse_obj(db.clubs.find_one({"cid": club_input["cid"]}))
-
-    return SimpleClubType.from_pydantic(updated_sample)
-
 
 """
 MEMBER MUTATIONS
@@ -267,7 +20,7 @@ def non_deleted_members(member_input) -> MemberType:
     Function to return non-deleted members for a particular cid, uid
     Only to be used in admin functions, as it returns both approved/non-approved members.
     """
-    updated_sample = db.members.find_one(
+    updated_sample = membersdb.find_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -319,7 +72,7 @@ def unique_roles_id(uid, cid):
             }
         }
     ]
-    db.members.update_one(
+    membersdb.update_one(
         {
             "$and": [
                 {"cid": cid},
@@ -346,7 +99,7 @@ def createMember(memberInput: FullMemberInput, info: Info) -> MemberType:
     if (member_input["cid"] != uid or user["role"] != "club") and user["role"] != "cc":
         raise Exception("Not Authenticated to access this API")
 
-    if db.members.find_one(
+    if membersdb.find_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -378,17 +131,14 @@ def createMember(memberInput: FullMemberInput, info: Info) -> MemberType:
     member_input["roles"] = roles
 
     # DB STUFF
-    created_id = db.members.insert_one(member_input).inserted_id
+    created_id = membersdb.insert_one(member_input).inserted_id
     unique_roles_id(member_input["uid"], member_input["cid"])
 
     created_sample = Member.parse_obj(
-        db.members.find_one({"_id": created_id}, {"_id": 0})
+        membersdb.find_one({"_id": created_id}, {"_id": 0})
     )
 
     return MemberType.from_pydantic(created_sample)
-
-
-# TODO: Add check for end_year >= start_year
 
 
 @strawberry.mutation
@@ -426,7 +176,7 @@ def editMember(memberInput: FullMemberInput, info: Info) -> MemberType:
         roles.append(role)
 
     # DB STUFF
-    db.members.update_one(
+    membersdb.update_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -456,7 +206,7 @@ def deleteMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
     if (member_input["cid"] != uid or user["role"] != "club") and user["role"] != "cc":
         raise Exception("Not Authenticated to access this API")
 
-    existing_data = db.members.find_one(
+    existing_data = membersdb.find_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -469,7 +219,7 @@ def deleteMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
         raise Exception("No such Record")
 
     if "rid" not in member_input or not member_input["rid"]:
-        db.members.delete_one(
+        membersdb.delete_one(
             {
                 "$and": [
                     {"cid": member_input["cid"]},
@@ -487,7 +237,7 @@ def deleteMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
         roles.append(i)
 
     # DB STUFF
-    db.members.update_one(
+    membersdb.update_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -516,7 +266,7 @@ def approveMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
     if user["role"] != "cc":
         raise Exception("Not Authenticated to access this API")
 
-    existing_data = db.members.find_one(
+    existing_data = membersdb.find_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -539,7 +289,7 @@ def approveMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
         roles.append(i)
 
     # DB STUFF
-    db.members.update_one(
+    membersdb.update_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -568,7 +318,7 @@ def rejectMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
     if user["role"] != "cc":
         raise Exception("Not Authenticated to access this API")
 
-    existing_data = db.members.find_one(
+    existing_data = membersdb.find_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -591,7 +341,7 @@ def rejectMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
         roles.append(i)
 
     # DB STUFF
-    db.members.update_one(
+    membersdb.update_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -619,7 +369,7 @@ def rejectMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
 #     if member_input["cid"] != uid and role != "club":
 #         raise Exception("Not Authenticated to access this API")
 
-#     created_id = db.clubs.update_one(
+#     created_id = clubsdb.update_one(
 #         {
 #             "$and": [
 #                 {"cid": member_input["cid"]},
@@ -631,16 +381,12 @@ def rejectMember(memberInput: SimpleMemberInput, info: Info) -> MemberType:
 #         {"$set": {"end_year": datetime.now().year}},
 #     )
 
-#     created_sample = Member.parse_obj(db.members.find_one({"_id": created_id}))
+#     created_sample = Member.parse_obj(membersdb.find_one({"_id": created_id}))
 #     return MemberType.from_pydantic(created_sample)
 
 
 # register all mutations
 mutations = [
-    createClub,
-    editClub,
-    deleteClub,
-    restartClub,
     createMember,
     editMember,
     deleteMember,
