+++ queries_members.py
@@ -3,110 +3,14 @@ import strawberry
 from fastapi.encoders import jsonable_encoder
 from typing import List
 
-from db import db
+from db import membersdb
 
 # import all models and types
 from otypes import Info
 
-from models import Club, Member
-from otypes import SimpleClubType, FullClubType, SimpleClubInput, SimpleMemberInput
+from models import Member
+from otypes import SimpleClubInput, SimpleMemberInput
 from otypes import MemberType
-
-"""
-Club Queries
-"""
-
-
-# fetch all active clubs
-@strawberry.field
-def activeClubs(info: Info) -> List[SimpleClubType]:
-    """
-    Description: Returns all the currently active clubs.
-    Scope: Public
-    Return Type: List[SimpleClubType]
-    Input: None
-    """
-    results = db.clubs.find({"state": "active"}, {"_id": 0})
-    clubs = [SimpleClubType.from_pydantic(Club.parse_obj(result)) for result in results]
-
-    return clubs
-
-
-@strawberry.field
-def allClubs(info: Info) -> List[SimpleClubType]:
-    """
-    Description:
-        For CC:
-            Returns all the currently active/deleted clubs.
-        For Public:
-            Returns all the currently active clubs.
-    Scope: CC (For All Clubs), Public (For Active Clubs)
-    Return Type: List[SimpleClubType]
-    Input: None
-    """
-    user = info.context.user
-    if user is None:
-        role = "public"
-    else:
-        role = user["role"]
-
-    results = []
-    if role in ["cc"]:
-        results = db.clubs.find()
-    else:
-        results = db.clubs.find({"state": "active"}, {"_id": 0})
-
-    clubs = []
-    for result in results:
-        clubs.append(SimpleClubType.from_pydantic(Club.parse_obj(result)))
-
-    return clubs
-
-
-@strawberry.field
-def club(clubInput: SimpleClubInput, info: Info) -> FullClubType:
-    """
-    Description: Returns complete details of the club, from its club-id.
-        For CC:
-            Returns details even for deleted clubs.
-        For Public:
-            Returns details only for the active clubs.
-    Scope: Public
-    Return Type: FullClubType
-    Input: SimpleClubInput (cid)
-    Errors:
-        - cid doesn't exist
-        - if not cc and club is deleted
-    """
-    user = info.context.user
-    club_input = jsonable_encoder(clubInput)
-
-    result = None
-    club = db.clubs.find_one({"cid": club_input["cid"].lower()}, {"_id": 0})
-
-    if not club:
-        raise Exception("No Club Found")
-
-    # check if club is deleted
-    if club["state"] == "deleted":
-        # if deleted, check if requesting user is admin
-        if (user is not None) and (user["role"] in ["cc"]):
-            result = Club.parse_obj(club)
-
-        # if not admin, raise error
-        else:
-            raise Exception("No Club Found")
-
-    # if not deleted, return club
-    else:
-        result = Club.parse_obj(club)
-
-    if result:
-        return FullClubType.from_pydantic(result)
-
-    else:
-        raise Exception("No Club Result Found")
-
 
 """
 Member Queries
@@ -132,7 +36,7 @@ def member(memberInput: SimpleMemberInput, info: Info) -> MemberType:
     if (member_input["cid"] != uid or user["role"] != "club") and user["role"] != "cc":
         raise Exception("Not Authenticated to access this API")
 
-    member = db.members.find_one(
+    member = membersdb.find_one(
         {
             "$and": [
                 {"cid": member_input["cid"]},
@@ -162,7 +66,7 @@ def memberRoles(uid: str, info: Info) -> List[MemberType]:
     else:
         role = user["role"]
 
-    results = db.members.find({"uid": uid}, {"_id": 0})
+    results = membersdb.find({"uid": uid}, {"_id": 0})
 
     if not results:
         raise Exception("No Member Result/s Found")
@@ -210,9 +114,9 @@ def members(clubInput: SimpleClubInput, info: Info) -> List[MemberType]:
     club_input = jsonable_encoder(clubInput)
 
     if role not in ["cc"] or club_input["cid"] != "clubs":
-        results = db.members.find({"cid": club_input["cid"]}, {"_id": 0})
+        results = membersdb.find({"cid": club_input["cid"]}, {"_id": 0})
     else:
-        results = db.members.find({}, {"_id": 0})
+        results = membersdb.find({}, {"_id": 0})
 
     if results:
         members = []
@@ -264,9 +168,9 @@ def currentMembers(clubInput: SimpleClubInput, info: Info) -> List[MemberType]:
         if role != "cc":
             raise Exception("Not Authenticated")
 
-        results = db.members.find({}, {"_id": 0})
+        results = membersdb.find({}, {"_id": 0})
     else:
-        results = db.members.find({"cid": club_input["cid"]}, {"_id": 0})
+        results = membersdb.find({"cid": club_input["cid"]}, {"_id": 0})
 
     if results:
         members = []
@@ -302,7 +206,7 @@ def pendingMembers(info: Info) -> List[MemberType]:
     if user is None or user["role"] not in ["cc"]:
         raise Exception("Not Authenticated")
 
-    results = db.members.find({}, {"_id": 0})
+    results = membersdb.find({}, {"_id": 0})
 
     if results:
         members = []
@@ -326,9 +230,6 @@ def pendingMembers(info: Info) -> List[MemberType]:
 
 # register all queries
 queries = [
-    activeClubs,
-    allClubs,
-    club,
     member,
     memberRoles,
     members,
